Fixed bug (incomplete migration to BigDecimal) that was preventing the estimated standard routing costs of a product to be computed.
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@777125 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/manufacturing/src/org/ofbiz/manufacturing/routing/RoutingServices.java
@@ -74,10 +74,10 @@ public class RoutingServices {
         long estimatedTaskTime = ProductionRun.getEstimatedTaskTime(task, quantity, productId, routingId, dispatcher);
         result.put("estimatedTaskTime", Long.valueOf(estimatedTaskTime));
         if (task != null && task.get("estimatedSetupMillis") != null) {
-            result.put("setupTime", task.getDouble("estimatedSetupMillis"));
+            result.put("setupTime", task.getBigDecimal("estimatedSetupMillis"));
         }
         if (task != null && task.get("estimatedMilliSeconds") != null) {
-            result.put("taskUnitTime", task.getDouble("estimatedMilliSeconds"));
+            result.put("taskUnitTime", task.getBigDecimal("estimatedMilliSeconds"));
         }
         return result;
     }
